// src/modules/consumer/type/message-catalog.type.ts
export interface MessageProductCatalog {
  id: string;
  title: string;
  owner: string;
  category: string;
  price: number;
  description: string;
  type: string;
}

export interface MessageCategoryCatalog {
  id: string;
  title: string;
  owner: string;
  description: string;
  type: string;
}